Revert 7f0c59d5c1291c087eccc4fc70014190de6a439f

+++ build.sh
@@ -88,22 +88,24 @@ run_stage(){
 	if [ -f ${STAGE_DIR}/EXPORT_IMAGE ]; then
 		EXPORT_DIRS="${EXPORT_DIRS} ${STAGE_DIR}"
 	fi
-	if [ "${CLEAN}" = "1" ]; then
-		if [ -d ${ROOTFS_DIR} ]; then
-			rm -rf ${ROOTFS_DIR}
+	if [ ! -f SKIP ]; then
+		if [ "${CLEAN}" = "1" ]; then
+			if [ -d ${ROOTFS_DIR} ]; then
+				rm -rf ${ROOTFS_DIR}
+			fi
 		fi
-	fi
-	if [ -x prerun.sh ]; then
-		log "Begin ${STAGE_DIR}/prerun.sh"
-		./prerun.sh
-		log "End ${STAGE_DIR}/prerun.sh"
-	fi
-	for SUB_STAGE_DIR in ${STAGE_DIR}/*; do
-		if [ -d ${SUB_STAGE_DIR} ] &&
-			[ ! -f ${SUB_STAGE_DIR}/SKIP ]; then
-			run_sub_stage
+		if [ -x prerun.sh ]; then
+			log "Begin ${STAGE_DIR}/prerun.sh"
+			./prerun.sh
+			log "End ${STAGE_DIR}/prerun.sh"
 		fi
-	done
+		for SUB_STAGE_DIR in ${STAGE_DIR}/*; do
+			if [ -d ${SUB_STAGE_DIR} ] &&
+			   [ ! -f ${SUB_STAGE_DIR}/SKIP ]; then
+				run_sub_stage
+			fi
+		done
+	fi
 	unmount ${WORK_DIR}/${STAGE}
 	PREV_STAGE=${STAGE}
 	PREV_STAGE_DIR=${STAGE_DIR}
@@ -124,12 +126,6 @@ fi
 if [ -z "${IMG_NAME}" ]; then
 	echo "IMG_NAME not set" 1>&2
 	exit 1
-fi
-
-if [ -n "${RUN_STAGE}" ]; then
-	echo "Running ONLY stage${RUN_STAGE}"
-elif [ -n "${MAX_STAGE}" ]; then
-	echo "Running stage${MAX_STAGE} build"
 fi
 
 export IMG_DATE=${IMG_DATE:-"$(date -u +%Y-%m-%d)"}
@@ -170,12 +166,7 @@ mkdir -p ${WORK_DIR}
 log "Begin ${BASE_DIR}"
 
 for STAGE_DIR in ${BASE_DIR}/stage*; do
-	STAGE_DIR_NUM=$(echo $STAGE_DIR | grep -o -E "[0-9]+$")
-	if [[ (-z $RUN_STAGE || $STAGE_DIR_NUM -eq $RUN_STAGE) && (-z $MAX_STAGE || $STAGE_DIR_NUM -le $MAX_STAGE) ]]; then
-		run_stage
-	else
-		echo "Skipping ${STAGE_DIR}"
-	fi
+	run_stage
 done
 
 CLEAN=1
